feat: added ActiveCall view to the React Native SDK
